+++ kramer/src/main/java/com/chiralbehaviors/layout/AutoLayout.java
@@ -20,7 +20,6 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 
 import com.chiralbehaviors.layout.Layout.LayoutModel;
-import com.chiralbehaviors.layout.control.JsonControl;
 import com.chiralbehaviors.layout.schema.Relation;
 import com.chiralbehaviors.layout.schema.SchemaNode;
 import com.fasterxml.jackson.databind.JsonNode;
