+++ kramer/src/main/java/com/chiralbehaviors/layout/Outline.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.chiralbehaviors.layout.control;
+package com.chiralbehaviors.layout;
 
 import java.util.ArrayList;
 import java.util.Collection;
