+++ kramer/src/main/java/com/chiralbehaviors/layout/schema/Primitive.java
@@ -21,8 +21,8 @@ import java.util.function.Consumer;
 import java.util.function.Function;
 
 import com.chiralbehaviors.layout.Layout;
+import com.chiralbehaviors.layout.NestedTable;
 import com.chiralbehaviors.layout.Layout.PrimitiveLayout;
-import com.chiralbehaviors.layout.control.NestedTable;
 import com.fasterxml.jackson.databind.JsonNode;
 
 import javafx.scene.Parent;
