+++ kramer/src/main/java/com/chiralbehaviors/layout/schema/Relation.java
@@ -22,10 +22,10 @@ import java.util.List;
 import java.util.function.Consumer;
 import java.util.function.Function;
 
+import com.chiralbehaviors.layout.JsonControl;
 import com.chiralbehaviors.layout.Layout;
+import com.chiralbehaviors.layout.NestedTable;
 import com.chiralbehaviors.layout.Layout.RelationLayout;
-import com.chiralbehaviors.layout.control.JsonControl;
-import com.chiralbehaviors.layout.control.NestedTable;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ArrayNode;
