+++ kramer/src/main/java/com/chiralbehaviors/layout/schema/SchemaNode.java
@@ -23,8 +23,8 @@ import java.util.function.Consumer;
 import java.util.function.Function;
 
 import com.chiralbehaviors.layout.Layout;
+import com.chiralbehaviors.layout.NestedTable;
 import com.chiralbehaviors.layout.Layout.SchemaNodeLayout;
-import com.chiralbehaviors.layout.control.NestedTable;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.JsonNodeFactory;
